app: Set as entrypoint, implement limit on cycles run, remove nomask
Got rid of main.py as this is not an appropriate entrypoint (it is
outside of the cvtransit package). Limit number of cycles run. Add tests,
which were previously missing.

Removed nomask, which was previously intended to determine the chances
that a commuter would not wear a mask. It was not implemented anywhere in
the app, might reinstate later.